+++ BenHinkleRecipes.DataAccessLayer/DAL/Repos/UserFavoriteRepository.cs
@@ -1,9 +1,8 @@
-﻿using BenHinkleRecipes.DAL.Context;
-using BenHinkleRecipes.Interfaces.RepoInterfaces;
-using BenHinkleRecipes.Models.RepoModels;
-using Microsoft.EntityFrameworkCore;
+﻿using BenHinkleRecipes.DataAccessLayer.Context;
+using BenHinkleRecipes.DataAccessLayer.Interfaces.RepoInterfaces;
+using BenHinkleRecipes.DataAccessLayer.Models.RepoModels;
 
-namespace BenHinkleRecipes.DAL.Repos
+namespace BenHinkleRecipes.DataAccessLayer.Repos
 {
     public class UserFavoriteRepository : IUserFavoriteRepository
     {
