+++ BenHinkleRecipes.DataAccessLayer/Models/RepoModels/RecipeRepoModel.cs
@@ -1,18 +1,15 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace BenHinkleRecipes.Models.RepoModels
+namespace BenHinkleRecipes.DataAccessLayer.Models.RepoModels
 {
-    public class UserRecipeRepoModel
+    public class RecipeRepoModel
     {
-        [Key]
-        public int UserRecipe_ID { get; set; }
-        public int recipe_id { get; set; }
+        public int Id { get; set; }
         public string? RecipeName { get; set; }
         public string? Description { get; set; }
         [MaxLength]
         public byte[]? RecipePhotoFront { get; set; }
         [MaxLength]
         public byte[]? RecipePhotoBack { get; set; }
-        public string userName { get; set; }
     }
 }
