+++ BenHinkleRecipes.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BenHinkleRecipes", "BenHinkleRecipes\BenHinkleRecipes.csproj", "{2B8E0DE6-DF3A-415A-AAAD-636653E6B7D2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BenHinkleRecipes", "BenHinkleRecipes\BenHinkleRecipes.csproj", "{2B8E0DE6-DF3A-415A-AAAD-636653E6B7D2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BenHinkleRecipes.DataAccessLayer", "BenHinkleRecipes.DataAccessLayer\BenHinkleRecipes.DataAccessLayer.csproj", "{F5D8F1D4-3DC9-4D31-8ACE-B9ACA61A0DFC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{2B8E0DE6-DF3A-415A-AAAD-636653E6B7D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2B8E0DE6-DF3A-415A-AAAD-636653E6B7D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2B8E0DE6-DF3A-415A-AAAD-636653E6B7D2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F5D8F1D4-3DC9-4D31-8ACE-B9ACA61A0DFC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F5D8F1D4-3DC9-4D31-8ACE-B9ACA61A0DFC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F5D8F1D4-3DC9-4D31-8ACE-B9ACA61A0DFC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F5D8F1D4-3DC9-4D31-8ACE-B9ACA61A0DFC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
