+++ BenHinkleRecipes/BenHinkleRecipes.csproj
@@ -26,4 +26,8 @@
     <PackageReference Include="System.Composition.TypedParts" Version="6.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\BenHinkleRecipes.DataAccessLayer\BenHinkleRecipes.DataAccessLayer.csproj" />
+  </ItemGroup>
+
 </Project>
