+++ BenHinkleRecipes/Interfaces/ServiceInterfaces/IIngredientService.cs
@@ -1,4 +1,4 @@
-﻿using BenHinkleRecipes.Models.RepoModels;
+﻿using BenHinkleRecipes.DataAccessLayer.Models.RepoModels;
 using BenHinkleRecipes.Models.ViewModels;
 
 namespace BenHinkleRecipes.Interfaces.ServiceInterfaces
